+++ chain/chains.go
@@ -31,7 +31,8 @@ func createGenerator(trend string, pointer *bandwagon, pool *sync.WaitGroup) {
 
 	tweets := getTweets(trend)
 	dictionary := buildDictionary(tweets)
-	starters := findStarters(tweets)
+	starters := findWordsAtIndex(tweets, 0)
+	followers := findWordsAtIndex(tweets, 1)
 
 	// Gather tweets from a given trend
 	// Build an n-gram dictionary from the tweets
@@ -39,21 +40,27 @@ func createGenerator(trend string, pointer *bandwagon, pool *sync.WaitGroup) {
 	// (THIS IS THE HARD PART)
 	*pointer = func() string {
 		starter := selectFrom(starters)
-		tweet := []string{starter}
+		follower := selectFrom(followers)
+		tweet := []string{starter, " " + follower}
 		length := len(starter)
 		current := starter
 
 		for true {
-			candidates := dictionary[current]
-			choice := selectFrom(candidates)
+			if candidates, ok := dictionary[current]; ok {
+				choice := selectFrom(candidates)
 
-			if len(choice)+length > 240 {
+				if len(choice)+length > limit {
+					// This composition is longer than twitter will allow
+					break
+				}
+
+				tweet = append(tweet, " "+choice)
+				length += len(choice)
+				current = choice
+			} else {
+				// we have reached a n-gram that has never been followed before
 				break
 			}
-
-			tweet = append(tweet, " "+choice)
-			length += len(choice)
-			current = choice
 		}
 
 		composedTweet := join(tweet...)
@@ -90,23 +97,4 @@ func MakeChains() {
 
 	// reset what bandwagon we're currently on
 	currentBandwagons = upcomingBandwagons
-}
-
-// concatenate some strings
-func join(strs ...string) string {
-	var ret string
-	for _, str := range strs {
-		ret += str
-	}
-	return ret
-}
-
-//
-// TweetFromTrend will try and generate a tweet from the given trend
-//
-func TweetFromTrend(trend string) string {
-	if generator, ok := currentBandwagons[trend]; ok {
-		return join(generator(), "\n")
-	}
-	return join("\"", trend, "\" is a not a currently available bandwagon\n")
 }
